+++ src/cli/dev.ts
@@ -4,6 +4,7 @@ import express, { Express } from 'express'
 import { toRootAbsolute } from '../utils'
 import { createServer as _createServer } from 'vite'
 import { CYAN, GREEN, RESET } from '../constants'
+import { Render } from '../types'
 
 const createServer = async (
   root = process.cwd()
@@ -20,22 +21,29 @@ const createServer = async (
     }
   })
 
-  const { default: render } = await vite.ssrLoadModule(
+  const { default: render } = (await vite.ssrLoadModule(
     toRootAbsolute('src', 'entry-server')
-  )
+  )) as { default: Render }
   const index = readFileSync(toRootAbsolute('index.html'), 'utf-8')
   app.use(vite.middlewares)
 
   app.use('*', async ({ originalUrl }, res) => {
     const context = {} as { url: string }
-    const appHtml = await render(originalUrl, context)
+    const { bodyTags, bodyAttrs, htmlAttrs, headTags } = await render(
+      originalUrl,
+      {}
+    )
 
     if (context.url) {
       return res.redirect(301, context.url)
     }
 
     const template = await vite.transformIndexHtml(originalUrl, index)
-    const html = template.replace(`<!--app-html-->`, appHtml) ?? ''
+    const html = template
+      .replace('<html', `<html ${htmlAttrs ?? ''}`)
+      .replace('<body', `<body ${bodyAttrs ?? ''}`)
+      .replace('</head>', `${headTags ?? ''}\n</head>`)
+      .replace(`<!--app-html-->`, bodyTags)
 
     res.status(200).set({ 'Content-Type': 'text/html' }).end(html)
   })