+++ src/types.ts
@@ -1,0 +1,9 @@
+export type Render = (
+  url: string,
+  manifest: Record<string, string[] | undefined>
+) => Promise<{
+  bodyTags: string
+  headTags?: string
+  htmlAttrs?: string
+  bodyAttrs?: string
+}>